requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
charset-normalizer==2.1.1
click==8.1.3
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==0.4.3
pathspec==0.11.0
platformdirs==2.6.2
python-dotenv==0.21.1
tomli==2.0.1
yarl==1.8.2
